# Optimize range iteration startup performance